assure `conflicts::mark()` writes its file atomically
That way it can't be observed half-written, or remain in a half-written
state in case of crash.
