+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -21,6 +21,7 @@ gitbutler-id.workspace = true
 gitbutler-time.workspace = true
 gitbutler-commit.workspace = true
 gitbutler-url.workspace = true
+gitbutler-fs.workspace = true
 serde = { workspace = true, features = ["std"] }
 bstr = "1.9.1"
 diffy = "0.3.0"
