+++ crates/gitbutler-oplog/src/reflog.rs
@@ -2,7 +2,7 @@ use anyhow::{Context, Result};
 use gitbutler_branch::{
     GITBUTLER_INTEGRATION_COMMIT_AUTHOR_EMAIL, GITBUTLER_INTEGRATION_COMMIT_AUTHOR_NAME,
 };
-use gitbutler_fs::fs::write;
+use gitbutler_fs::write;
 use gix::config::tree::Key;
 use std::path::Path;
